% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaco.R
\docType{data}
\name{cjkextractja}
\alias{cjkextractja}
\alias{cjkextractzh}
\alias{cjkextractko}
\title{Personal name authority records}
\format{
cjkextractja is a \code{tbl_df} containing the following fields:
\describe{
\item{lccn}{LC control number, lower case 'n', no space}
\item{tag008_32}{Descriptive cataloging rule. Differentiated personal name when coded “a”}
\item{tag100_field}{Name in the 100 field}
\item{tag400_CJK}{Number of Chinese, Japanese or Korean references in 400 fields}
\item{tag400_other}{Number of other non-Latin references in 400 fields}
\item{tag400_roman}{Number of Latin references in 400 fields}
\item{rda}{Whether or not RDA record}
\item{tag670}{Number of 670 fields}
\item{tag667}{Has non-Latin script reference been evaluated?}
\item{Language}{Language of name record}
}

cjkextractzh is a \code{tbl_df} containing the following fields:
\describe{
\item{lccn}{LC control number, lower case 'n', no space}
\item{tag008_32}{Descriptive cataloging rule. Differentiated personal name when coded “a”}
\item{tag100_field}{Name in the 100 field}
\item{tag400_CJK}{Number of Chinese, Japanese or Korean references in 400 fields}
\item{tag400_other}{Number of other non-Latin references in 400 fields}
\item{tag400_roman}{Number of Latin references in 400 fields}
\item{rda}{Whether or not RDA record}
\item{tag670}{Number of 670 fields}
\item{tag667}{Has non-Latin script reference been evaluated?}
\item{notes}{If not CJK or minority name}
}

cjkextractko is a \code{tbl_df} containing the following fields:
\describe{
\item{lccn}{LC control number, lower case 'n', no space}
\item{tag008_32}{Descriptive cataloging rule. Differentiated personal name when coded “a”}
\item{tag100_field}{Name in the 100 field}
\item{tag400_CJK}{Number of Chinese, Japanese or Korean references in 400 fields}
\item{tag400_other}{Number of other non-Latin references in 400 fields}
\item{tag400_roman}{Number of Latin references in 400 fields}
\item{rda}{Whether or not RDA record}
\item{tag670}{Number of 670 fields}
\item{tag667}{Has non-Latin script reference been evaluated?}
\item{tag400Hangul}{Number of Hangul references in 400 fields}
\item{Hangul_only}{}
\item{note}{}
}
}
\source{
The data comes from the NACO CJK Funnel References Project.
}
\description{
The \href{https://www.loc.gov/aba/pcc/naco/CJK/NACO-CJK-Funnel-References-Project-Guidelines.docx}{NACO CJK Funnel References Project Guidelines}\link{@naco2019naco}
refer to three Google spreadsheets containing personal name authorities for use
in the funnel project. These spreadsheets derive from 268,128 records extracted by
Gary Strawn. While they emphasize Chinese, Japanese and Korean, when coupled
with VIAF they access a rich resource of name authority records from
the various VIAF contributors in various languages. Please refer to the
\href{https://www.loc.gov/aba/publications/FreeDCM/DCM/Z01.pdf}{Descriptive Cataloging Manual}
for detailed information on the tag fields. For example, the tag009_32 and tag670
columns relate to compliance with RDA name authority records.
}
\details{
Personal name authority records
}
\seealso{
\link{rnaco}
\link{marcNotes}
\link{marctables}
\link{unihan_variants}
\link{nacotransform}
\link{cjkextracts}
}
