% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaco.R
\docType{data}
\name{jctPrecedence}
\alias{jctPrecedence}
\alias{kyoikuKanji}
\alias{twXiaoxueWenzi}
\alias{zhXiaoxueWenzi}
\alias{unihanPrecedenceData}
\title{Unicode code point precedence tables based on kUnihanCore2020 source tags and educational grade levels}
\format{
\code{jctPrecedence} is a \code{tbl_df} containing the following fields:
\describe{
\item{ucs}{Unicode code point of ideograph}
\item{kTotalStrokes}{Total strokes in ideograph}
\item{jCoreRank}{Ranking based on kUnihanCore2020 source tags favoring 'J' (Japan)}
\item{jpRank}{Ranking based on educational grade level of ideograph in Japan}
\item{gCoreRank}{Ranking based on kUnihanCore2020 source tags favoring 'G' (PRC)}
\item{zhRank}{Ranking based on educational grade level of ideograph in PRC}
\item{tCoreRank}{Ranking based on kUnihanCore2020 source tags favoring 'T' (Taiwan)}
\item{twRank}{Ranking based on educational grade level of ideograph in Taiwan}
}

kyoikuKanji is a \code{tbl_df} containing the following fields:
\describe{
\item{ucs}{Unicode code point of ideograph}
\item{char}{Ideograph}
\item{jpGrade}{Grade level (1-6)}
}

twXiaoxueWenzi is a \code{tbl_df} containing the following fields:
\describe{
\item{ucs}{Unicode code point of ideograph}
\item{char}{Ideograph}
\item{twGrade}{Grade level (1-9)}
}

zhXiaoxueWenzi is a \code{tbl_df} containing the following fields:
\describe{
\item{ucs}{Unicode code point of ideograph}
\item{char}{Ideograph}
\item{zhGrade}{Grade level (1-6)}
}

\code{unihanPrecedenceData} is a \code{tbl_df} containing the following fields:
\describe{
\item{ucs}{Unicode code point of ideograph}
\item{kFrequency}{Unihan property: A rough frequency measurement for the ideograph based on analysis of traditional Chinese USENET postings; ideographs with a kFrequency of 1 are the most common, those with a kFrequency of 2 are less common, and so on, through a kFrequency of 5.}
\item{kTotalStrokes}{Unihan property: The total number of strokes in the ideograph (including the radical). }
\item{kIICore}{Unihan property: Used for ideographs which are in IICore, the IRG-produced minimal set of required ideographs for East Asian use.}
\item{kUnihanCore2020}{Unihan property: Used for ideographs which are in the Unihan Core 2020 set.}
\item{kJinmeiyoKanji}{Unihan property: The year that corresponds to the Jinmei-yō Kanji (人名用漢字) table in which the ideograph appears.}
\item{kJis0}{Unihan property: The JIS X 0208-1990 mapping for this ideograph in row-cell form.}
\item{kJis1}{Unihan property: The JIS X 0212-1990 mapping for this ideograph in row-cell form.}
\item{kJIS0213}{Unihan property: The JIS X 0213:2004 mapping for this ideograph in plane-row-cell form.}
\item{kJoyoKanji}{Unihan property: The year that corresponds to the Jōyō Kanji (常用漢字) table in which the ideograph appears.}
\item{kMojiJoho}{Unihan property: mappings from the Moji Jōhō Kiban database by CITPC (Character Information Technology Promotion Council) }
}
}
\source{
The data for \code{kyoikuKanji} comes from Japan's Ministry of Education, Culture, Sports, Science and Technology (MEXT);
for \code{twXiaoxueWenzi} from \href{http://www.microweiidv.com/weipage/n9gradeword.htm}; and for \code{zhXiaoxueWenzi} from
\href{https://cd.jiajiaoban.com/e/20210314/604de778686f4.shtml}.
The data for ```unihanPrecedenceData```` comes from the XML Representation of Unicode 15.1.0 UCD flat file.
}
\description{
Detailed description of the data
}
\seealso{
\link{rnaco}
\href{../doc/unihanprecedencedata.html}{\code{vignette("unihanprecedencedata", package = "rnaco")}}
\href{../doc/kyoikuKanji.html}{\code{vignette("kyoikuKanji", package = "rnaco")}}
\href{../doc/xiaoxuewenziTW.html}{\code{vignette("xiaoxuewenziTW", package = "rnaco")}}
\href{../doc/xiaoxuewenziZH.html}{\code{vignette("xiaoxuewenziZH", package = "rnaco")}}
}
