% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaco.R
\docType{data}
\name{jztTranslitRules}
\alias{jztTranslitRules}
\title{Descriptive data title}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{ucs}{}
\item{membership}{}
\item{from}{}
\item{to.JA}{}
\item{rule.JA}{}
\item{to.ZH}{}
\item{rule.ZH}{}
\item{to.TW}{}
\item{rule.TW}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{rnaco}
}
