% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaco.R
\docType{data}
\name{marcComponentInputMethodCharacters}
\alias{marcComponentInputMethodCharacters}
\title{Component Input Method Characters code table (February 6, 2003)}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{marc}{MARC-8 code (in hex) for the character as coming from the G0 graphic set}
\item{ucs}{UCS/Unicode code point (hexmode)}
\item{utf8}{UTF-8 code (in hex) for the UCS character}
\item{rep}{representation of the character (where possible)}
\item{isAlt}{For some tables alternate encodings in Unicode and UTF-8 are given.
When that occurs the alternate Unicode and alternate UTF-8 columns follow the character name.}
\item{isCombining}{}
\item{name}{MARC character name, followed by the UCS name}
}
}
\source{
The data comes from the \href{https://www.loc.gov/marc/specifications/codetables.xml}{Library of Congress}.
}
\description{
This table contains 35 mappings of encodings for component input method characters from the East Asian Coded Character set (ANSI/NISO Z39.64, or "EACC") to character encodings in the Private Use Area (PUA) of the Universal Character Set (UCS, ISO-IEC 10646)/Unicode. Character codes are given in hexadecimal notation. Each character is presented on a separate row.
}
\details{
Detailed description of the data
}
\seealso{
\link{rnaco}
\link{marcBasicArabic}
\link{marcBasicCyrillic}
\link{marcBasicGreek}
\link{marcBasicHebrew}
\link{marcEastAsianIdeographsHan}
\link{marcEastAsianPunctuationMarks}
\link{marcExtendedArabic}
\link{marcExtendedCyrillic}
\link{marcExtendedLatinANSEL}
\link{marcGreekSymbols}
\link{marcJapaneseHiraganaandKatakana}
\link{marcKoreanHangul}
\link{marcSubscripts}
\link{marcSuperscripts}
\link{unihan_variants}
}
