% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaco.R
\docType{data}
\name{marcNotes}
\alias{marcNotes}
\title{Descriptive data title}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{marcName}{}
\item{marcPath}{}
\item{marcNote}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{rnaco}
}
