% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaco.R
\docType{data}
\name{marcSuperscripts}
\alias{marcSuperscripts}
\title{Descriptive data title}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{marc}{}
\item{ucs}{}
\item{utf8}{}
\item{rep}{}
\item{isAlt}{}
\item{isCombining}{}
\item{name}{}
}
}
\source{
The data comes from________________________.
}
\description{
Detailed description of the data
}
\seealso{
\link{rnaco}
\link{marcBasicArabic}
\link{marcBasicCyrillic}
\link{marcBasicGreek}
\link{marcBasicHebrew}
\link{marcBasicLatinASCII}
\link{marcComponentInputMethodCharacters}
\link{marcEastAsianIdeographsHan}
\link{marcEastAsianPunctuationMarks}
\link{marcExtendedArabic}
\link{marcExtendedCyrillic}
\link{marcExtendedLatinANSEL}
\link{marcGreekSymbols}
\link{marcJapaneseHiraganaandKatakana}
\link{marcKoreanHangul}
\link{marcSubscripts}
}
