% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaco.R
\docType{data}
\name{marcBasicLatinASCII}
\alias{marcBasicLatinASCII}
\alias{marcBasicArabic}
\alias{marcBasicCyrillic}
\alias{marcBasicGreek}
\alias{marcBasicHebrew}
\alias{marcComponentInputMethodCharacters}
\alias{marcEastAsianIdeographsHan}
\alias{marcEastAsianPunctuationMarks}
\alias{marcExtendedArabic}
\alias{marcExtendedCyrillic}
\alias{marcExtendedLatinANSEL}
\alias{marcGreekSymbols}
\alias{marcJapaneseHiraganaandKatakana}
\alias{marcKoreanHangul}
\alias{marcSubscripts}
\alias{marcSuperscripts}
\title{Basic Latin (ASCII) code table (January 2000) ISOcode=42}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{marc}{MARC-8 code (in hex) for the character as coming from the G0 graphic set}
\item{ucs}{UCS/Unicode code point (hexmode)}
\item{utf8}{UTF-8 code (in hex) for the UCS character}
\item{rep}{representation of the character (where possible)}
\item{isAlt}{For some tables alternate encodings in Unicode and UTF-8 are given.
When that occurs the alternate Unicode and alternate UTF-8 columns follow the character name.}
\item{isCombining}{}
\item{name}{MARC character name, followed by the UCS name}
}

a \code{tbl_df} containing the following fields:
\describe{
\item{marc}{MARC-8 code (in hex) for the character as coming from the G0 graphic set}
\item{ucs}{UCS/Unicode code point (hexmode)}
\item{utf8}{UTF-8 code (in hex) for the UCS character}
\item{rep}{representation of the character (where possible)}
\item{isAlt}{For some tables alternate encodings in Unicode and UTF-8 are given.
When that occurs the alternate Unicode and alternate UTF-8 columns follow the character name.}
\item{isCombining}{}
\item{name}{MARC character name, followed by the UCS name}
}

a \code{tbl_df} containing the following fields:
\describe{
\item{marc}{MARC-8 code (in hex) for the character as coming from the G0 graphic set}
\item{ucs}{UCS/Unicode code point (hexmode)}
\item{utf8}{UTF-8 code (in hex) for the UCS character}
\item{rep}{representation of the character (where possible)}
\item{isAlt}{For some tables alternate encodings in Unicode and UTF-8 are given.
When that occurs the alternate Unicode and alternate UTF-8 columns follow the character name.}
\item{isCombining}{}
\item{name}{MARC character name, followed by the UCS name}
}

a \code{tbl_df} containing the following fields:
\describe{
\item{marc}{MARC-8 code (in hex) for the character as coming from the G0 graphic set}
\item{ucs}{UCS/Unicode code point (hexmode)}
\item{utf8}{UTF-8 code (in hex) for the UCS character}
\item{rep}{representation of the character (where possible)}
\item{isAlt}{For some tables alternate encodings in Unicode and UTF-8 are given.
When that occurs the alternate Unicode and alternate UTF-8 columns follow the character name.}
\item{isCombining}{}
\item{name}{MARC character name, followed by the UCS name}
}

a \code{tbl_df} containing the following fields:
\describe{
\item{marc}{MARC-8 code (in hex) for the character as coming from the G0 graphic set}
\item{ucs}{UCS/Unicode code point (hexmode)}
\item{utf8}{UTF-8 code (in hex) for the UCS character}
\item{rep}{representation of the character (where possible)}
\item{isAlt}{For some tables alternate encodings in Unicode and UTF-8 are given.
When that occurs the alternate Unicode and alternate UTF-8 columns follow the character name.}
\item{isCombining}{}
\item{name}{MARC character name, followed by the UCS name}
}

a \code{tbl_df} containing the following fields:
\describe{
\item{marc}{MARC-8 code (in hex) for the character as coming from the G0 graphic set}
\item{ucs}{UCS/Unicode code point (hexmode)}
\item{utf8}{UTF-8 code (in hex) for the UCS character}
\item{rep}{representation of the character (where possible)}
\item{isAlt}{For some tables alternate encodings in Unicode and UTF-8 are given.
When that occurs the alternate Unicode and alternate UTF-8 columns follow the character name.}
\item{isCombining}{}
\item{name}{MARC character name, followed by the UCS name}
}

a \code{tbl_df} containing the following fields:
\describe{
\item{marc}{MARC-8 code (in hex) for the character as coming from the G0 graphic set}
\item{ucs}{UCS/Unicode code point (hexmode)}
\item{utf8}{UTF-8 code (in hex) for the UCS character}
\item{rep}{representation of the character (where possible)}
\item{isAlt}{For some tables alternate encodings in Unicode and UTF-8 are given.
When that occurs the alternate Unicode and alternate UTF-8 columns follow the character name.}
\item{isCombining}{}
\item{name}{MARC character name, followed by the UCS name}
}
}
\source{
The data comes from the \href{https://www.loc.gov/marc/specifications/codetables.xml}{Library of Congress}.
}
\description{
From the Basic and Extended Latin character sets

From the Basic and Extended Arabic character sets

From the Basic and Extended Cyrillic character sets

From the Greek character set

From the Basic Hebrew character set

This table contains 35 mappings of encodings for component input method characters from the East Asian Coded Character set (ANSI/NISO Z39.64, or "EACC") to character encodings in the Private Use Area (PUA) of the Universal Character Set (UCS, ISO-IEC 10646)/Unicode. Character codes are given in hexadecimal notation. Each character is presented on a separate row.

This table contains 13,478 mappings of character encodings for East Asian ideographs from the East Asian Coded Character set (ANSI/NISO Z39.64, or "EACC", including 10 "Version J" additions that did not appear in the published standard) to character encodings in the Universal Character Set (UCS, ISO-IEC 10646)/Unicode. Character codes are given in hexadecimal notation. Each character is presented on a separate row.

This table contains 25 mappings of character encodings for East Asian punctuation marks from the East Asian Coded Character set (ANSI/NISO Z39.64, or "EACC") to character encodings in the Universal Character Set (UCS, ISO-IEC 10646)/Unicode. Character codes are given in hexadecimal notation.

From the Basic and Extended Arabic character sets

From the Basic and Extended Cyrillic character sets

From the Extended Latin (ANSEL) character sets

Greek Symbols code table, ISO code 67 (January 2000, Updated September 2004)

This table contains 172 mappings of character encodings for Japanese Hiragana and Katakana from the East Asian Coded Character set (ANSI/NISO Z39.64, or "EACC") to character encodings in the Universal Character Set (UCS, ISO-IEC 10646)/Unicode.

This table contains 2,028 mappings of character encodings for Korean hangul from the East Asian Coded Character set (ANSI/NISO Z39.64, or "EACC") to character encodings in the Universal Character Set (UCS, ISO-IEC 10646)/Unicode.

Subscripts code table, ISO code 62 (January 2000)

Superscripts code table, ISO code 70 (January 2000)
}
\details{
Detailed description of the data

See the vignette \code{vignette("marctables")}.

Detailed description of the data

Detailed description of the data

Detailed description of the data

Detailed description of the data

Detailed description of the data

Detailed description of the data

Detailed description of the data

Detailed description of the data

Detailed description of the data

Detailed description of the data

Detailed description of the data

Detailed description of the data

Detailed description of the data

Detailed description of the data

Detailed description of the data
}
\seealso{
\link{rnaco}
\link{marcNotes}
\link{marctables}
\link{unihan_variants}
\link{nacotransform}
\link{cjkextracts}
}
