% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moji-transform.R
\name{moji_transform}
\alias{moji_transform}
\title{Apply normalization based on CITPC MJ Map}
\usage{
moji_transform(s, encoding="UTF-8")
}
\arguments{
\item{s}{String}
}
\value{
Transformed (normalized) string
}
\description{
Transform (normalize) a string for VIAF comparisons.
}
