% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaco.R
\docType{data}
\name{mojiMapdf}
\alias{mojiMapdf}
\title{Descriptive data title}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{mjcode}{}
\item{from}{}
\item{to}{}
\item{relation}{}
}
}
\source{
The data comes from the Japanese Character Information Technology Promotion Council. \link{https://moji.or.jp/mojikiban/}
}
\description{
Detailed description of the data
}
\seealso{
\link{rnaco}
\link{mojiVariants}
\link{unihanVariants}
}
