% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaco.R
\docType{data}
\name{mojiVariants}
\alias{mojiVariants}
\title{Descriptive data title}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{ucsindex}{}
\item{ucsvariant}{}
\item{membership}{}
\item{from}{}
\item{to}{}
\item{rule}{}
}
}
\source{
The data comes from the Japanese Character Information Technology Promotion Council. \link{https://moji.or.jp/mojikiban/}
}
\description{
Detailed description of the data
}
\seealso{
\link{rnaco}
\link{mojiMapdf}
\link{unihanVariants}
\href{../doc/citpcmojivariants.html}{\code{vignette("citpcmojivariants", package = "rnaco")}}
}
