% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiform-transform.R
\name{multiform_transform}
\alias{multiform_transform}
\title{Apply normalization based on CITPC MJ Map}
\usage{
multiform_transform(s, form=c("JA","ZH","TW"), encoding="UTF-8")
}
\arguments{
\item{s}{String}

\item{form}{string selected from "JA","ZH" and "TW"}
}
\value{
Transformed (normalized) string
}
\description{
Transform (normalize) a string for VIAF comparisons.
}
