% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naco-transform.R
\name{naco_transform}
\alias{naco_transform}
\title{Apply NACO normalization (Authority File Comparison Rules)}
\usage{
naco_transform(s, firstcomma=FALSE, encoding="UTF-8")
}
\arguments{
\item{firstcomma}{Retain the first comma in the string}

\item{encoding}{UTF-8 only}

\item{s}{String}
}
\value{
Transformed (normalized) string
}
\description{
Transform (normalize) a string for according to
LOC's Authority file comparison rules.
}
\details{
'naco_transform()' applies the authority file comparison
rules (NACO normalization) to a string
}
\seealso{
\itemize{
\item Please see \code{vignette("naco", package = "rnaco")} for a detailed walk-through of
the NACO authority file comparison rules.
}
}
