% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-transliteration-rules.R, R/rnaco.R
\docType{data}
\name{nacotransform}
\alias{nacotransform}
\alias{get_transliteration_rules}
\alias{unconditionalMappings}
\alias{numericrules}
\alias{step6rulestring}
\alias{step7lltranslitrule}
\title{Get transiteration rules}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{ucs}{UCS/Unicode code point (hexmode)}
\item{lower}{Lower case mapping}
\item{title}{Title case mapping}
\item{upper}{Upper case mapping}
\item{name}{UCS name}
}

numericrules is a \code{tbl_df} containing the following fields:
\describe{
\item{V1}{chr: code point}
\item{V2}{chr: name of code point}
\item{V3}{chr: Unicode category}
\item{V8}{int: Integer replacement for code point}
\item{from}{chr: Character for transliteration rule}
\item{to}{chr: Character for transliteration rule}
\item{rule}{chr: Transliteration rule}
}

step6rulestring is in the form of a \code{string}

step7lltranslitrule is in the form of a \code{string}
}
\source{
The data comes from the \href{https://www.unicode.org/versions/Unicode15.1.0/}{Unicode® Standard}
provides a file UCD.zip that contains the file SpecialCasings.txt.
}
\usage{
get_transliteration_rules(s, encoding="UTF-8")
}
\arguments{
\item{s}{String}

\item{encoding}{UTF-8 only}
}
\value{
character string containing ICU transliteration rules for
input string
}
\description{
Transform (normalize) a string for according to
LOC's Authority file comparison rules.

The 2009 version of the authority file comparison rules call
for using the unconditional mappings from the
Unicode® Standard.
These mappings provide lower, title and upper case mapping for characters such as
the German ß (U+00DF, Latin small letter sharp S).
The mappings used
for this package are found in the file SpecialCasings.txt, which is contained in the
file UCD.zip

Detailed description of the data
}
\seealso{
\link{rnaco}
\link{marcNotes}
\link{marctables}
\link{unihanVariants}
\link{nacotransform}
\link{cjkextracts}
\href{../doc/naco.html}{\code{vignette("naco", package = "rnaco")}}
\href{../doc/nacoDatasets.html}{\code{vignette("nacoDatasets", package = "rnaco")}}
}
