% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaco.R
\docType{package}
\name{rnaco}
\alias{rnaco}
\alias{rnaco-package}
\title{DName Normalization for the VIAF API}
\description{
This package combines, in one place, data and functions for
normalizing transforms of names and titles in VIAF records. Please
see the vignettes for details.
}
\details{
rnaco
A data package for rnaco.

Use \code{data(package='rnaco')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{marcNotes}
\link{marctables}
\link{mojiMapdf}
\link{mojiVariants}
\link{unihanVariants}
\link{nacotransform}
\link{cjkextracts}
}
\author{
\strong{Maintainer}: Alan Engel \email{alan@ttcrossroads.net} (\href{https://orcid.org/0000-0003-0070-2347}{ORCID})

}
