% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster-diagnostics.R
\name{show_cluster}
\alias{show_cluster}
\title{Apply normalization based on CITPC MJ Map}
\usage{
show_cluster(chr, communities)
}
\arguments{
\item{chr}{A character}

\item{graph}{An igraph graph object, corresponding to communities}

\item{communities}{A communities object, the result of an igraph community detection function
such as \code{\link[igraph:cluster_walktrap]{igraph::cluster_walktrap()}}.}
}
\description{
Transform (normalize) a string for VIAF comparisons.
}
