% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaco.R
\docType{data}
\name{unihan_variants}
\alias{unihan_variants}
\title{Unihan CJK variants}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{ucsindex}{UCS/Unicode code point (hexmode)}
\item{varianttype}{Type of variant, e.g., kSemanticVariant}
\item{ucsvariant}{UCS/Unicode code point (hexmode)}
}
}
\source{
The data comes from the \href{https://www.unicode.org/reports/tr38/}{UNICODE HAN DATABASE}\link{@jenkins2021uax}
provides a file Unihan.zip that contains a file Unihan\_Variants.txt.
}
\description{
Variants loaded from the Unicode Han Database. Lists of
variants have been unnested so that there is only one character and one variant per line.
This is at the expense of having multiple rows for ucsindexes that have multiple variants.
}
\details{
Detailed description of the data
}
\seealso{
\link{rnaco}
\link{marcBasicArabic}
\link{marcBasicCyrillic}
\link{marcBasicGreek}
\link{marcBasicHebrew}
\link{marcComponentInputMethodCharacters}
\link{marcEastAsianIdeographsHan}
\link{marcEastAsianPunctuationMarks}
\link{marcExtendedArabic}
\link{marcExtendedCyrillic}
\link{marcExtendedLatinANSEL}
\link{marcGreekSymbols}
\link{marcJapaneseHiraganaandKatakana}
\link{marcKoreanHangul}
\link{marcSubscripts}
\link{unconditionalMappings}
}
