% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viaf-transform.R
\name{viaf_transform}
\alias{viaf_transform}
\title{Data frame (tibble) from xml}
\usage{
viaf_transform(s, mode=c("name","title"), encoding="UTF-8")
}
\arguments{
\item{s}{String}

\item{mode}{"title" mode removes all spaces. "name" mode leaves single
spaces between name parts.}
}
\value{
Transformed (normalized) string
}
\description{
Transform (normalize) a string for VIAF comparisons.
}
