% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml2df.R
\name{xml2df}
\alias{xml2df}
\title{Data frame (tibble) from xml}
\usage{
xml2df(x)
}
\arguments{
\item{x}{Loaded xml document.}
}
\value{
The desired tibble.
}
\description{
'xml2df()' reads a specially structured xml, transforms it to an
xml, and from that xml produces a tibble.

An example of the special structure is below.
The names of the elements under the rows will become the
column names. The columns will have the type given by
the 'ct' attribute.

\preformatted{
<marcEastAsianIdeographsHan>
<row>
<marc ct="character">212F30</marc>
<ucs ct="character">0x3007</ucs>
<utf8 ct="character">E38087</utf8>
<isAlt ct="logical"/>
<isCombining ct="logical"/>
<name ct="character">East Asian ideograph (number zero)</name>
</row>
  ...
</marcEastAsianIdeographsHan>
}
}
