% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaco.R
\docType{data}
\name{cjkextractja}
\alias{cjkextractja}
\alias{cjkextractzh}
\title{Personal name authority records}
\format{
cjkextractja is a \code{tbl_df} containing the following fields:
\describe{
\item{010 $a}{}
\item{008/32}{Descriptive cataloging rule. Differentiated personal name when coded “a”}
\item{100 field}{Name in the 100 field}
\item{400 CJK}{Number of Chinese, Japanese or Korean references in 400 fields}
\item{400 other}{Number of other non-Latin references in 400 fields}
\item{400 roman}{Number of Latin references in 400 fields}
\item{rda}{Whether or not RDA record}
\item{670}{Number of 670 fields}
\item{667}{Has non-Latin script reference been evaluated?}
\item{Language}{Language of name record}
}

cjkextractzh is a \code{tbl_df} containing the following fields:
\describe{
\item{010 $a}{}
\item{008/32}{Descriptive cataloging rule. Differentiated personal name when coded “a”}
\item{100 field}{Name in the 100 field}
\item{400 CJK}{Number of Chinese, Japanese or Korean references in 400 fields}
\item{400 other}{Number of other non-Latin references in 400 fields}
\item{400 roman}{Number of Latin references in 400 fields}
\item{rda}{Whether or not RDA record}
\item{670}{Number of 670 fields}
\item{667}{Has non-Latin script reference been evaluated?}
\item{notes}{If not CJK or minority name}
}
}
\source{
The data comes from the NACO CJK Funnel References Project.
}
\description{
The \href{https://www.loc.gov/aba/pcc/naco/CJK/NACO-CJK-Funnel-References-Project-Guidelines.docx}{NACO CJK Funnel References Project Guidelines}\link{@naco2019naco}
refer to three Excel files containing personal name authorities for use
in the funnel project. These files contain 268,128 records extracted by
Gary Strawn. While they emphasize Chinese, Japanese and Korean, when coupled
with VIAF, they access a rich resource of name authority records from
the various VIAF contributors.
}
\details{
Personal name authority records
}
\seealso{
\link{rnaco}
\link{cjkextractzh}
}
