% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaco.R
\docType{data}
\name{jctPrecedence}
\alias{jctPrecedence}
\alias{kyoikuKanji}
\alias{twXiaoxueWenzi}
\alias{zhXiaoxueWenzi}
\alias{unihanPrecedenceData}
\title{Unicode code point precedence tables based on kUnihanCore2020 source tags and educational grade levels}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{ucs}{Unicode code point of ideograph}
\item{kTotalStrokes}{Total strokes in ideograph}
\item{jCoreRank}{Ranking based on kUnihanCore2020 source tags favoring 'J' (Japan)}
\item{jpRank}{Ranking based on educational grade level of ideograph in Japan}
\item{gCoreRank}{Ranking based on kUnihanCore2020 source tags favoring 'G' (PRC)}
\item{zhRank}{Ranking based on educational grade level of ideograph in PRC}
\item{tCoreRank}{Ranking based on kUnihanCore2020 source tags favoring 'T' (Taiwan)}
\item{twRank}{Ranking based on educational grade level of ideograph in Taiwan}
}

kyoikuKanji is a \code{tbl_df} containing the following fields:
\describe{
\item{ucs}{Unicode code point of ideograph}
\item{char}{Ideograph}
\item{jpGrade}{Grade level (1-6)}
}

twXiaoxueWenzi is a \code{tbl_df} containing the following fields:
\describe{
\item{ucs}{Unicode code point of ideograph}
\item{char}{Ideograph}
\item{twGrade}{Grade level (1-9)}
}

zhXiaoxueWenzi is a \code{tbl_df} containing the following fields:
\describe{
\item{ucs}{Unicode code point of ideograph}
\item{char}{Ideograph}
\item{zhGrade}{Grade level (1-6)}
}

a \code{tbl_df} containing the following fields:
\describe{
\item{ucs}{}
\item{kFrequency}{}
\item{kTotalStrokes}{}
\item{kIICore}{}
\item{kUnihanCore2020}{}
\item{kJinmeiyoKanji}{}
\item{kJis0}{}
\item{kJis1}{}
\item{kJIS0213}{}
\item{kJoyoKanji}{}
\item{kMojiJoho}{}
}
}
\source{
The data comes from Japan's Ministry of Education, Culture, Sports, Science and Technology (MEXT);
\href{http://www.microweiidv.com/weipage/n9gradeword.htm}; and
\href{https://cd.jiajiaoban.com/e/20210314/604de778686f4.shtml}.

The data comes from the XML Representation of Unicode 15.1.0 UCD flat file.
}
\description{
Detailed description of the data

Detailed description of the data
}
\seealso{
\link{rnaco}
\href{../doc/unihanprecedencedata.html}{\code{vignette("unihanprecedencedata", package = "rnaco")}}
\href{../doc/kyoikuKanji.html}{\code{vignette("kyoikuKanji", package = "rnaco")}}
\href{../doc/xiaoxuewenziTW.html}{\code{vignette("xiaoxuewenziTW", package = "rnaco")}}
\href{../doc/xiaoxuewenziZH.html}{\code{vignette("xiaoxuewenziZH", package = "rnaco")}}

\link{rnaco}
}
