% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaco.R
\docType{data}
\name{marcNotes}
\alias{marcNotes}
\title{Names, XPaths and notes for MARC-8 code tables}
\format{
a \code{tbl_df} containing the following fields:
\describe{
\item{marcName}{name of code table}
\item{marcPath}{XPath location of table in source xml}
\item{marcNote}{notes from source xml}
}
}
\source{
The data comes from the \href{https://www.loc.gov/marc/specifications/codetables.xml}{Library of Congress}.
}
\description{
The source xml file, codetables.xml, contains
named character sets and groupings of character sets together
informative notes. This dataset contains the names, XPath locations
of the character sets or groupings, and the notes pertaining to
the respective sets or groupings.
}
\details{
Detailed description of the data
}
\seealso{
\link{rnaco}
\link{marcNotes}
\link{marctables}
\link{unihanVariants}
\link{nacotransform}
\link{cjkextracts}
\href{../doc/marctables.html}{\code{vignette("marctables", package = "rnaco")}}
}
