% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnaco.R
\docType{package}
\name{rnaco}
\alias{rnaco}
\alias{rnaco-package}
\title{Package Title}
\description{
A description of the data package
}
\details{
rnaco
A data package for rnaco.

Use \code{data(package='rnaco')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{cjkextractko}
}
\author{
\strong{Maintainer}: Alan Engel \email{alan@ttcrossroads.net} (\href{https://orcid.org/0000-0003-0070-2347}{ORCID})

}
